add a serial connector to allow custom path/boudrate for serial connection or auto mode. refactor sourcecode into folders
