+++ src/lib.rs
@@ -1,10 +1,4 @@
-pub mod action;
-pub mod file;
+pub mod models;
 pub mod printer_actions;
-pub mod temperature;
 
-mod pre_send_handler;
-mod response_handler;
 mod serial;
-mod temp_interval;
-
