+++ src/models/mod.rs
@@ -1,0 +1,4 @@
+pub mod action;
+pub mod file;
+pub mod serial_connector;
+pub mod temperature;
