+++ src/models/serial_connector.rs
@@ -1,0 +1,6 @@
+pub enum SerialConnector {
+    // automatically connect to first serial port found
+    Auto,
+    // manually specify (serialport, boudrate)
+    Manual(String, u32),
+}
